Fix errors

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	long long int longlongintType;
 	float floatType;
 
-	printf("Size of a char: %zu bytes\n", sizeof(charType));
-	printf("Size of a int: %zu bytes\n", sizeof(intType));
-	printf("Size of a long int: %zu bytes\n", sizeof(longintType));
-	printf("Size of a long long int: %zu bytes\n," sizeof(longlongintType));
-	printf("Size of a float: %zu bytes\n," sizeof(floatType));
+	printf("Size of a char: %lu bytes\n", (unsigned long)sizeof(charType));
+	printf("Size of a int: %lu bytes\n", (unsigned long)sizeof(intType));
+	printf("Size of a long int: %lu bytes\n", (unsigned long)sizeof(longintType));
+	printf("Size of a long long int: %lu bytes\n," (unsigned long)sizeof(longlongintType));
+	printf("Size of a float: %lu bytes\n," (unsigned long)sizeof(floatType));
 
 	return(0)
 }
